crewai/philosophical_ai/pyproject.toml
[project]
name = "philosophical-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.148.0",
    "langchain-groq>=0.3.6",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
]
